requirements.txt
liboqs-python==0.14.1
matplotlib>=3.5.0
pandas>=1.3.0
tqdm>=4.60.0
jupyter>=1.0.0
ipykernel>=6.0.0
asn1crypto>=1.5.1

